pyproject.toml
[tool.poetry]
name = "espn-fantasy-telegram-bot"
version = "0.1.0"
description = "A Telegram bot that provides ESPN fantasy league updates including scores, standings, waivers, and news"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "espn_fantasy_bot"}]

[tool.poetry.dependencies]
python = "^3.8.1"
python-telegram-bot = {extras = ["job-queue"], version = "^20.7"}
espn-api = "^0.45.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
schedule = "^1.2.0"

[tool.poetry.group.dev.dependencies]
# Development dependencies removed for production deployment

[tool.poetry.scripts]
espn-bot = "espn_fantasy_bot:main"
espn-fantasy-cli = "cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Development tool configurations removed for production deployment
